Testcases runned

+++ testcases/MyPower/MyPower.java
@@ -1,6 +1,6 @@
 class MyPower {
-    public static double power(double x, int n) {
-        double res = 1.0;
+    public static int power(int x, int n) {
+        int res = 1;
         for(int i = 0; i < n; i++)
 	    res = res * x;
         return res;
